pyproject.toml
[build-system]
requires = ["setuptools", "gwerks", "pytest"]
build-backend = "setuptools.build_meta"

#[build-system]
#requires = ["hatchling", "gwerks", "pytest"]
#build-backend = "hatchling.build"

[project]
name = "fwq"
authors = [{name = "Ron Glasmann", email = "ron@glasmann.net"}]
readme = "README.md"
license = { text = "MIT" }
dynamic = ["version"]
dependencies = [
    "gwerks>=25.3.18",
    "pyyaml",
    "greenstalk",
    "requests",
    "uvicorn"
]

[project.optional-dependencies]
aws = ["boto3"]

[project.scripts]
fwq = "fwq.cli:cli_fwq"

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.dynamic]
version = {attr = "fwq.__version__"}

[tool.pytest.ini_options]
testpaths = ["src-tests"]
addopts = [
    "--import-mode=importlib",
]